* EModel: all defined model attributes are initializing in constructor; * db: fixed bug in db_insert() when insert data has key "id" with "null" value; + basic support for cacheable templates. Template file should contain ".chacheable" in its name; * file_cache: bug fixing & refactoring; * xml: fixed typo; * page.htm: prints company name if defined instead of app name in footer; + money: show_money() formats numeric value as number_format($value, 2, ",", " ");

+++ src/functions/file_cache.php
@@ -2,13 +2,13 @@
 // require PHP 5.4
 define("FILE_CACHE_ENABLED", true);
 define("FILE_CACHE_DELETE_FLAG", uniqid("file_cache_del"));
-define("FILE_CACHE_DIR", FILES_DIR.".file_cache/");
+define("FILE_CACHE_DIR", DATA_DIR.".cache/file_cache/");
 define("FILE_CACHE_TTL_DEFAULT", 3600); // Time to live in seconds
 function file_cache($value = null){
         
     if ( ! FILE_CACHE_ENABLED ) return null;
     
-    $key = _cache_key();
+    $key = glog_codify(_cache_key(), GLOG_CODIFY_FILENAME);
             
     if ( is_null($value) ){  // get from cache 
         return _file_cache($key);
@@ -36,19 +36,30 @@ function file_cache_del($key){
     file_cache_set($key, FILE_CACHE_DELETE_FLAG);
 }
 function file_cache_get($key){
-    return _file_cache($key);
+    $hash = glog_codify($key, GLOG_CODIFY_FILENAME);
+    if ($hash !== $key){
+        dosyslog(__FUNCTION__.get_callee().": WARNING: key:'".$key."' is not suitable for file name. Consider to change it.");
+    }
+    return _file_cache($hash);
 };
 function file_cache_set($key, $value){
-    return _file_cache($key, $value);
+    $hash = glog_codify($key, GLOG_CODIFY_FILENAME);
+    if ($hash !== $key){
+        dosyslog(__FUNCTION__.get_callee().": WARNING: key:'".$key."' is not suitable for file name. Consider to change it.");
+    }
+    return _file_cache($hash, $value);
 }
 function file_cached($key = null){
     
     if ( ! FILE_CACHE_ENABLED ) return null;
         
     if ( is_null($key) ){
-        $hash = __cache_kay();
+        $hash = glog_codify(_cache_key(), GLOG_CODIFY_FILENAME);
     }else{
-        $hash = $key;
+        $hash = glog_codify($key, GLOG_CODIFY_FILENAME);
+        if ($hash !== $key){
+            dosyslog(__FUNCTION__.get_callee().": WARNING: key:'".$key."' is not suitable for file name. Consider to change it.");
+        }
     }
     
     return ! is_null(_file_cache($hash));
@@ -72,30 +83,41 @@ function _file_cache($key, $value = null){
     if ( is_null($value) ){  // get from cache 
         if ( isset($file_cache[$key]) ){
             
-            // check TTL
-            if (time() - filemtime(FILE_CACHE_DIR . $key) <= get_file_cache_ttl() ){
-                dosyslog(__FUNCTION__.get_callee().": DEBUG: Попадание: ".$key.".");
-                $file_cache[$key]++;
-                return glog_file_read( FILE_CACHE_DIR . $key );
+            if (file_exists(FILE_CACHE_DIR . $key)){
+                // check TTL
+                if (time() - filemtime(FILE_CACHE_DIR . $key) <= get_file_cache_ttl() ){
+                    dosyslog(__FUNCTION__.get_callee().": DEBUG: Попадание: ".$key.".");
+                    $file_cache[$key]++;
+                    return glog_file_read( FILE_CACHE_DIR . $key );
+                }else{
+                    unlink(FILE_CACHE_DIR . $key);
+                    dosyslog(__FUNCTION__.get_callee().": DEBUG: Инвалидация: ".$key." после ". get_file_cache_ttl() ." ceк.");
+                    return null;
+                };
             }else{
-                unlink(FILE_CACHE_DIR . $key);
-                dosyslog(__FUNCTION__.get_callee().": DEBUG: Инвалидация: ".$key." после ". get_file_cache_ttl() ." ceк.");
+                dosyslog(__FUNCTION__.get_callee().": DEBUG: Предполагалась инвалидация: ".$key." после ". get_file_cache_ttl() ." ceк., но файл уже не существует.");
                 return null;
-            };
+            }
         }else{
             dosyslog(__FUNCTION__.get_callee().": DEBUG: Промах: ".$key.".");
             return null;
         };
-    }else{                     // set into cache
-        if ($value === FILE_CACHE_DELETE_FLAG){
+    }else{
+    
+        if ($value === FILE_CACHE_DELETE_FLAG){ // delete from cache
+        
             if (isset($file_cache[$key])){
                 unlink(FILE_CACHE_DIR . $key);
-            };
+                dosyslog(__FUNCTION__.get_callee().": DEBUG: Принудительная инвалидация: ".$key);
+            }else{
+                dosyslog(__FUNCTION__.get_callee().": WARNING: Ошибка ключ '".$key."' отсутствует в кеше.");
+            }
             return;
         }
         
+        // set into cache
         if (file_put_contents(FILE_CACHE_DIR . $key, $value)){
-            $file_cache[$key] = 0;
+            $file_cache[$key] = 0; // value is a counter for key have been accessed/requested.
             dosyslog(__FUNCTION__.get_callee().": DEBUG: Запись: ".$key.".");
         }else{
             dosyslog(__FUNCTION__.get_callee().": ERROR: Can not write cache file '".FILE_CACHE_DIR . $key.". Data lost.");
